requirements.txt
pandas
tqdm
networkx
matplotlib.pyplot